backend/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
psycopg2-binary==2.9.9
sqlparse==0.5.0
requests==2.31.0
typing_extensions==4.11.0